requirements.txt
scrapy>=0.24.4
mysql-python
python-firebase
-e git://github.com/saymedia/seosuite.git#egg=seosuite
